requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-environ==0.8.1
psycopg2==2.9.3
sqlparse==0.4.2

django-modeltranslation~=0.18.2
pytils~=0.4.1